requirements.txt
HTMLTestRunner-Python3==0.8.0
Appium-Python-Client==1.1.0
